+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for polysemy-extra
 
+## v0.1.3.0
+
+* Add `raise4Under`, `rotateEffects4L`, `rotateEffects4R`.
+* Add `reverseEffects2` ,`reverseEffects3`. `reverseEffects4`.
+
 ## v0.1.2.0
 
 * Add `rotateEffects2`, `rotateEffects3L` and `rotateEffects3R`.